deleating foto from carpet storage in local

+++ app/Http/Controllers/EmpleadoController.php
@@ -88,7 +88,10 @@ class EmpleadoController extends Controller
     public function destroy($id)
     {
         //
-        Empleado::destroy($id);
+        $empleado= Empleado::findOrFail($id);
+        if( Storage::delete('public/'.$empleado->foto)){
+            Empleado::destroy($id);
+        }
         return redirect('/empleado');
     }
 }
